rewrite cleaner LM optimizer
